Added NextPVR log backup button and status reporting
Added a button to allow the user to backup NextPVR's log files into a separate timestamped folder with a single click, just in case they need to backup the logs for another reason.

Added code that changes the text in the title bar of the app to indicate when it's waiting to start monitoring, and when it's in the process of closing down for whatever reason.