Fixed several potential errors.
Changed the code so that the service monitoring thread is the one to clean up the resources, just in case the form is closed during the initial waiting period, so that the resources persist until all threads finish. Before this, the service monitoring thread would attempt to use the log writer that had already been closed by the GUI thread.

Added code to handle instances where the application is run on a computer that doesn't have NextPVR installed. Before this, the code didn't compensate for that and would instead end up monitoring the first service in the retrieved list instead of the intended service.