Moved UI strings to Resources.resx.
Moved all of the strings that appear in the Main Form GUI via code into the Resources.resx file in an attempt to make the application capable of being Globalized.

The log file, however, will remain in English because it's really only useful to me, the app developer. It may also prove useful to the developers of NextPVR, and they seem to work primarily in English as well.